refactor question search and pagination logic with new helper functions for both rawsql and prisma
