+++ src/app/api/questions/search/route.ts
@@ -1,8 +1,6 @@
-import { COMPANIES, Company } from "@/constants/companies";
-import { DIFFICULTIES, Difficulty } from "@/constants/difficulties";
-import { Topic, TOPICS } from "@/constants/topics";
 import { prismaLib } from "@/lib/prisma";
 import { get200Response, UnknownServerError } from "@/utils/api-responses";
+import { getPagination, getWhereClause } from "@/utils/search-helpers";
 import { Prisma, Question } from "@prisma/client";
 
 export async function GET(req: Request) {
@@ -10,8 +8,18 @@ export async function GET(req: Request) {
     const url = new URL(req.url);
     const search = url.searchParams.get("query");
 
-    const whereClause = getWhereClause(url);
-    const paginationSQL = getPaginationSQL(url);
+    const topics = url.searchParams.get("topics") || undefined;
+    const difficulty = url.searchParams.get("difficulty") || undefined;
+    const companies = url.searchParams.get("companies") || undefined;
+
+    const whereClause = getWhereClause(topics, difficulty, companies, true);
+
+    const cursor = parseInt(url.searchParams.get("cursor") || "0");
+    const take = parseInt(url.searchParams.get("take") || "20");
+    const skip = parseInt(url.searchParams.get("skip") || "0");
+    const sortBy = url.searchParams.get("sortBy") || "id";
+
+    const pagination = getPagination(cursor, take, skip, sortBy, true);
 
     let baseQuery = `
       SELECT 
@@ -40,7 +48,7 @@ export async function GET(req: Request) {
       baseQuery += ` AND ${whereClause}`;
     }
 
-    baseQuery += paginationSQL;
+    baseQuery += pagination;
 
     // use of any here is straight fronm Prisma documentation, Also this should be safe
     // https://www.prisma.io/docs/orm/prisma-client/using-raw-sql/raw-queries#safely-using-queryraw-and-executeraw-in-more-complex-scenarios
@@ -64,99 +72,4 @@ export async function GET(req: Request) {
     console.error("Error during question search:", error);
     return UnknownServerError;
   }
-}
-
-function getWhereClause(url: URL) {
-  const topics = url.searchParams.get("topics");
-  const difficulty = url.searchParams.get("difficulty");
-  const companies = url.searchParams.get("companies");
-
-  let whereClause = "";
-
-  // topics should be formatted as a space-separated string of topic keys
-  if (topics) {
-    let topicArray: (string | undefined)[] = topics
-      .split(" ")
-      .map((t) => TOPICS[t as Topic]);
-    topicArray = topicArray.filter((t) => t !== undefined);
-    if (topicArray.length !== 0) {
-      if (whereClause) {
-        whereClause += " AND ";
-      }
-      whereClause += `"topics" && ARRAY[${topicArray
-        .map((t) => `'${t}'`)
-        .join(",")}]::"Topic"[]`;
-    }
-  }
-
-  // difficulty should be formatted as a space-separated string of difficulty levels
-  if (difficulty) {
-    let difficultyArray: (number | undefined)[] = difficulty
-      .split(" ")
-      .map((d) => DIFFICULTIES[d as Difficulty]);
-    difficultyArray = difficultyArray.filter((d) => d !== undefined);
-    if (difficultyArray.length !== 0) {
-      if (whereClause) {
-        whereClause += " AND ";
-      }
-      whereClause += `"difficulty" IN (${difficultyArray
-        .map((d) => `'${d}'`)
-        .join(",")})`;
-    }
-  }
-
-  // companies should be formatted as a space-separated string of company keys
-  if (companies) {
-    let companyArray: (string | undefined)[] = companies
-      .split(" ")
-      .map((c) => COMPANIES[c as Company]);
-    companyArray = companyArray.filter((c) => c !== undefined);
-    if (companyArray.length !== 0) {
-      if (whereClause) {
-        whereClause += " AND ";
-      }
-      whereClause += `"companies" && ARRAY[${companyArray
-        .map((c) => `'${c}'`)
-        .join(",")}]::"Company"[]`;
-    }
-  }
-
-  return whereClause;
-}
-
-function getPaginationSQL(url: URL) {
-  const cursor = parseInt(url.searchParams.get("cursor") || "0");
-  const take = parseInt(url.searchParams.get("take") || "20");
-  const skip = parseInt(url.searchParams.get("skip") || "0");
-  const sortBy = url.searchParams.get("sortBy") || "id";
-
-  let cursorClause = "";
-  if (cursor !== 0) {
-    cursorClause = ` AND "id" > ${cursor}`;
-  }
-
-  let orderLimitOffset = "";
-  if (sortBy === "rank") {
-    orderLimitOffset += ` ORDER BY "rank" DESC`;
-  } else {
-    orderLimitOffset += ` ORDER BY "id" ASC`;
-  }
-
-  if (skip > 0) {
-    orderLimitOffset += ` OFFSET ${skip}`;
-  }
-  if (take < 0) {
-    orderLimitOffset += ` LIMIT ${-take}`;
-    if (cursor !== 0) {
-      cursorClause = ` AND "id" < ${cursor}`;
-    }
-    orderLimitOffset = orderLimitOffset.replace(
-      'ORDER BY "id" ASC',
-      'ORDER BY "id" DESC'
-    );
-  } else {
-    orderLimitOffset += ` LIMIT ${take}`;
-  }
-
-  return cursorClause + orderLimitOffset;
 }
